# swagger.yaml
openapi: 3.0.0
info:
  title: ArtCollectibles
  contact: {}
  version: '1.0'
servers:
- url: https://art-collectibles.onrender.com/api/v1
  variables: {}
paths:
...
  /carts:
    get:
      tags:
      - Customer
      summary: Get Cart
      operationId: GetCart
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
    post:
      tags:
      - Customer
      summary: Create Cart
      operationId: CreateCart
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreateCartRequest'
              - example:
                  artCollectibleId: 63c88fc41841882bb80528bc
                  quantity: '1'
            example:
              artCollectibleId: 63c88fc41841882bb80528bc
              quantity: '1'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
      security:
      - bearer: []
...
components:
  schemas:
...
    CreateCartRequest:
      title: CreateCartRequest
      required:
      - artCollectibleId
      - quantity
      type: object
      properties:
        artCollectibleId:
          type: string
        quantity:
          type: string
      example:
        artCollectibleId: 63c88fc41841882bb80528bc
        quantity: '1'
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
security: []
...